// src/apis/proto_new/UserService/Commands/user_auth_command.ts
export interface LoginRequest {
    /**
     * @generated from protobuf field: google.protobuf.StringValue Tel = 1
     */
    tel?: StringValue;
    /**
     * @generated from protobuf field: google.protobuf.StringValue Password = 2
     */
    password?: StringValue;
}
/**
 * @generated from protobuf message UserAuth.LogoutRequest
 */
export interface LogoutRequest {
}
/**
 * @generated from protobuf message UserAuth.RefreshTokenRequest
 */
export interface RefreshTokenRequest {
    /**
     * @generated from protobuf field: optional google.protobuf.StringValue RefreshToken = 1
     */
    refreshToken?: StringValue;
}
/**
 * @generated from protobuf message UserAuth.RegisterRequest
 */
export interface RegisterRequest {
    /**
     * @generated from protobuf field: optional google.protobuf.StringValue Tel = 1
     */
    tel?: StringValue;
    /**
     * @generated from protobuf field: optional google.protobuf.StringValue Password = 2
     */
    password?: StringValue;
    /**
     * @generated from protobuf field: optional google.protobuf.BoolValue Type = 3
     */
    type?: BoolValue;
    /**
     * @generated from protobuf field: optional google.protobuf.StringValue FirstName = 4
     */
    firstName?: StringValue;
    /**
     * @generated from protobuf field: optional google.protobuf.StringValue LastName = 5
     */
    lastName?: StringValue;
    /**
     * @generated from protobuf field: optional google.protobuf.StringValue Email = 6
     */
    email?: StringValue;
    /**
     * @generated from protobuf field: optional google.protobuf.StringValue SendOTPMethod = 7
     */
    sendOTPMethod?: StringValue; // Currently support: EMAIL. Future support: SMS
    /**
     * @generated from protobuf field: optional google.protobuf.StringValue MailLanguage = 8
     */
    mailLanguage?: StringValue; // Currently support: VN (default), EN
}
/**
 * @generated from protobuf message UserAuth.ActivateAccountRequest
 */
export interface ActivateAccountRequest {
    /**
     * @generated from protobuf field: optional google.protobuf.StringValue Tel = 1
     */
    tel?: StringValue;
    /**
     * @generated from protobuf field: optional google.protobuf.StringValue OTP = 2
     */
    oTP?: StringValue;
}
/**
 * @generated from protobuf message UserAuth.VerifyOTPRequest
 */
export interface VerifyOTPRequest {
    /**
     * @generated from protobuf field: optional google.protobuf.StringValue Tel = 1
     */
    tel?: StringValue;
    /**
     * @generated from protobuf field: optional google.protobuf.StringValue OTP = 2
     */
    oTP?: StringValue;
}
/**
 * @generated from protobuf message UserAuth.ResetPasswordRequest
 */
export interface ResetPasswordRequest {
    /**
     * @generated from protobuf field: optional google.protobuf.StringValue Tel = 1
     */
    tel?: StringValue;
    /**
     * @generated from protobuf field: optional google.protobuf.StringValue Method = 2
     */
    method?: StringValue; // Currently support: EMAIL. Future support: SMS
    /**
     * @generated from protobuf field: optional google.protobuf.StringValue OTP = 3
     */
    oTP?: StringValue; // Need OTP send before to verify reset request is valid
    /**
     * @generated from protobuf field: optional google.protobuf.StringValue MailLanguage = 4
     */
    mailLanguage?: StringValue; // Currently support: VN (default), EN
}
/**
 * @generated from protobuf message UserAuth.ReSendOTPRequest
 */
export interface ReSendOTPRequest {
    /**
     * @generated from protobuf field: optional google.protobuf.StringValue Tel = 1
     */
    tel?: StringValue;
    /**
     * @generated from protobuf field: optional google.protobuf.StringValue Case = 2
     */
    case?: StringValue; // Use for chose appropriate mail content and subject for each case. Currently support: DEFAULT (default), ACTIVATE_ACCOUNT, RESET_PASSWORD
    /**
     * @generated from protobuf field: optional google.protobuf.StringValue Method = 3
     */
    method?: StringValue; // Currently support: EMAIL. Future support: SMS
    /**
     * @generated from protobuf field: optional google.protobuf.StringValue MailLanguage = 4
     */
    mailLanguage?: StringValue; // Currently support: VN (default), EN
    /**
     * @generated from protobuf field: optional google.protobuf.StringValue Email = 5
     */
    email?: StringValue;
}
/**
 * @generated from protobuf message UserAuth.GetSessionRequest
 */
export interface GetSessionRequest {
}
// @generated message type with reflection information, may provide speed optimized methods
class LoginRequest$Type extends MessageType<LoginRequest> {
    constructor() {
        super("UserAuth.LoginRequest", [
            { no: 1, name: "Tel", kind: "message", jsonName: "Tel", T: () => StringValue },
            { no: 2, name: "Password", kind: "message", jsonName: "Password", T: () => StringValue }
        ]);
    }
    create(value?: PartialMessage<LoginRequest>): LoginRequest {
        const message = globalThis.Object.create((this.messagePrototype!));
        if (value !== undefined)
            reflectionMergePartial<LoginRequest>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: LoginRequest): LoginRequest {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* google.protobuf.StringValue Tel */ 1:
                    message.tel = StringValue.internalBinaryRead(reader, reader.uint32(), options, message.tel);
                    break;
                case /* google.protobuf.StringValue Password */ 2:
                    message.password = StringValue.internalBinaryRead(reader, reader.uint32(), options, message.password);
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: LoginRequest, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* google.protobuf.StringValue Tel = 1; */
        if (message.tel)
            StringValue.internalBinaryWrite(message.tel, writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        /* google.protobuf.StringValue Password = 2; */
        if (message.password)
            StringValue.internalBinaryWrite(message.password, writer.tag(2, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message UserAuth.LoginRequest
 */
export const LoginRequest = new LoginRequest$Type();
// @generated message type with reflection information, may provide speed optimized methods
class LogoutRequest$Type extends MessageType<LogoutRequest> {
    constructor() {
        super("UserAuth.LogoutRequest", []);
    }
    create(value?: PartialMessage<LogoutRequest>): LogoutRequest {
        const message = globalThis.Object.create((this.messagePrototype!));
        if (value !== undefined)
            reflectionMergePartial<LogoutRequest>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: LogoutRequest): LogoutRequest {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: LogoutRequest, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message UserAuth.LogoutRequest
 */
export const LogoutRequest = new LogoutRequest$Type();
// ...
export const RefreshTokenRequest = new RefreshTokenRequest$Type();
// ...
export const RegisterRequest = new RegisterRequest$Type();
// @generated message type with reflection information, may provide speed optimized methods
class ActivateAccountRequest$Type extends MessageType<ActivateAccountRequest> {
    constructor() {
        super("UserAuth.ActivateAccountRequest", [
            { no: 1, name: "Tel", kind: "message", jsonName: "Tel", T: () => StringValue },
            { no: 2, name: "OTP", kind: "message", jsonName: "OTP", T: () => StringValue }
        ]);
    }
    create(value?: PartialMessage<ActivateAccountRequest>): ActivateAccountRequest {
        const message = globalThis.Object.create((this.messagePrototype!));
        if (value !== undefined)
            reflectionMergePartial<ActivateAccountRequest>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: ActivateAccountRequest): ActivateAccountRequest {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* optional google.protobuf.StringValue Tel */ 1:
                    message.tel = StringValue.internalBinaryRead(reader, reader.uint32(), options, message.tel);
                    break;
                case /* optional google.protobuf.StringValue OTP */ 2:
                    message.oTP = StringValue.internalBinaryRead(reader, reader.uint32(), options, message.oTP);
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: ActivateAccountRequest, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* optional google.protobuf.StringValue Tel = 1; */
        if (message.tel)
            StringValue.internalBinaryWrite(message.tel, writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        /* optional google.protobuf.StringValue OTP = 2; */
        if (message.oTP)
            StringValue.internalBinaryWrite(message.oTP, writer.tag(2, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message UserAuth.ActivateAccountRequest
 */
export const ActivateAccountRequest = new ActivateAccountRequest$Type();
// ...
export const VerifyOTPRequest = new VerifyOTPRequest$Type();
// ...
export const ResetPasswordRequest = new ResetPasswordRequest$Type();
// ...
export const ReSendOTPRequest = new ReSendOTPRequest$Type();